// src/app/core/models/user-databean.ts
import { UserExperience } from './user-experience';

export class UserDataBean {
    firstName: string;
    lastName: string;
    emailId: string;
    mobileNo: string;
    country: string;
    state: string;
    designation: string;
    city: string;
    languageKnown: string;
    userExperience: UserExperience[];
}